pyproject.toml
[project]
name = "legend-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "aioboto3>=13.0.0",
    "aiobotocore[boto3]>=2.15.0",
    "boto3>=1.35.0",
    "buildtools>=1.0.6",
    "deepgram-sdk>=3.0.0",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "onnxruntime>=1.23.2",
    "pipecat-ai>=0.0.98",
    "pipecat-ai-small-webrtc-prebuilt>=2.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "setuptools>=80.9.0",
    "vad>=1.0.2",
    "webrtcvad>=2.0.10",
    "webrtcvad-wheels>=2.0.14",
]
